feat: Add RMP scrapers and add dirs for different schools
