+++ .gitignore
@@ -1,0 +1,7 @@
+cache/
+output/
+
+node_modules/
+__pycache__/
+
+*.log
