+++ schools/elco/parser/pyproject.toml
@@ -1,0 +1,15 @@
+[tool.poetry]
+name = "scraper-thing"
+version = "0.1.0"
+description = ""
+authors = ["Edwin Kofler <edwin@kofler.dev>"]
+
+[tool.poetry.dependencies]
+python = "^3.10"
+beautifulsoup4 = "^4.11.1"
+
+[tool.poetry.dev-dependencies]
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
