+++ schools/elco/scraper/package.json
@@ -1,0 +1,13 @@
+{
+	"private": true,
+	"type": "module",
+	"scripts": {
+		"start": "node --loader ts-node/esm ./index.ts"
+	},
+	"devDependencies": {
+		"@types/node": "^17.0.34",
+		"puppeteer": "^14.1.0",
+		"ts-node": "^10.7.0",
+		"typescript": "^4.6.4"
+	}
+}
